requirements.txt
mistralai
python-dotenv
duckduckgo-search
markdown
xhtml2pdf
streamlit
fpdf2


